Unit Test SonarCloud Final

+++ src/main/java/co/edu/uniajc/estudiante/opemayfruitshop/model/Product.java
@@ -11,5 +11,5 @@ public class Product {
     private String id;
     private String name;
     private double price;
-    private int stock;
+    private double stock;
 }
